remove show error alert
