+++ wire-ios/Wire-iOS/Sources/UserInterface/UserProfile/ProfileViewController.swift
@@ -595,13 +595,31 @@ extension ProfileViewController: ProfileViewControllerViewModelDelegate {
     func presentError(_ error: LocalizedError) {
         typealias Strings = L10n.Localizable.Error.Connection
 
-        if let connectionError = error as? ConnectToUserError,
-           connectionError == .federationDenied {
-            let message = Strings.federationDeniedMessage(viewModel.user.name ?? "")
-            UIAlertController.showErrorAlert(title: "", message: message)
+        let title: String?
+        let message: String?
+        let style: UIAlertAction.Style
+
+        if let connectionError = error as? ConnectToUserError, connectionError == .federationDenied {
+            title = nil
+            message = Strings.federationDeniedMessage(viewModel.user.name ?? "")
+            style = .cancel
         } else {
-            presentLocalizedErrorAlert(error)
+            title = error.localizedDescription
+            message = error.failureReason
+            style = .default
         }
+
+        let alertController = UIAlertController(
+            title: title,
+            message: message,
+            preferredStyle: .alert
+        )
+        alertController.addAction(UIAlertAction(
+            title: L10n.Localizable.General.ok,
+            style: style
+        ))
+
+        present(alertController, animated: true)
     }
 
     func presentConversationCreationError(username: String) {
